% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx.R
\docType{data}
\name{fx}
\alias{fx}
\title{FX Test}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 270 rows and 13 columns.
}
\source{
\url{https://pandem2-test.clarisoft.com/}
}
\usage{
fx
}
\description{
FX Test
}
\keyword{datasets}
