% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid.R
\docType{data}
\name{owid}
\alias{owid}
\title{Our World in Data loaded on 19/11/2022}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 235693 rows and 67 columns.
}
\source{
\url{https://ourworldindata.org}
}
\usage{
owid
}
\description{
Our World in Data loaded on 19/11/2022
}
\keyword{datasets}
