% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth1.R
\name{synth1}
\alias{synth1}
\title{\code{synth1} Disaggregates a time series by sub-groups}
\usage{
synth1(ts_data, group_names, group_prob, setSeed = F, seedValue = 100)
}
\arguments{
\item{ts_data}{is the time series to be disaggregated}

\item{group_names}{the sub-group names}

\item{group_prob}{the probability of each sub-group appearing}
}
\value{
tibble with results
}
\description{
\code{synth1} Disaggregates a time series by sub-groups
}
