% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_ps.R
\name{synth_ps}
\alias{synth_ps}
\title{\code{synth_ps} 
Takes incidence and generates a noisy leading indicator to model
participatory surveillance}
\usage{
synth_ps(
  weekly_cases,
  dates,
  population_size,
  reporting_fraction = 0.5,
  proportion_enrolled = 0.0025,
  leading_time = 1,
  false_positive_fraction = 0.1,
  set_seed = F,
  SEED = 100,
  rbinomial_size = 5,
  active_users_lower = 0.5,
  active_users_upper = 0.9
)
}
\arguments{
\item{weekly_cases}{weekly reported cases}

\item{population_size}{population size}

\item{reporting_fraction}{estimated reporting fraction for cases}

\item{proportion_enrolled}{proportion enrolled in participatory surveillance}

\item{set_seed}{flag for setting the seed}

\item{SEED}{value for the seed}

\item{rbinomial_size}{dispersion parameter for negative binomial sample.}

\item{active_users_lower}{lower proportion of active users}

\item{active_users_higher}{higher proportion of active users}

\item{fp_fraction}{fraction false positives}
}
\value{
tibble with synthetic participatory surveillance numbers
}
\description{
\code{synth_ps} 
Takes incidence and generates a noisy leading indicator to model
participatory surveillance
}
