% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_variants.R
\name{synth_variants}
\alias{synth_variants}
\title{\code{synth_variants} 
Generates cases for five successive variants}
\usage{
synth_variants(
  cases_ts,
  time = length(cases_ts),
  percent_genomic_surveillance_min = 0.5,
  percent_genomic_surveillance_max = 0.8,
  variant_start_time_fractions = c(st_02 = 0.1, st_03 = 0.3, st_04 = 0.5, st_05 = 0.7),
  variant_growth_rates = c(gr_02 = 0.1, gr_03 = 0.2, gr_04 = 0.3, gr_05 = 0.05),
  variant_names = c(name_01 = "Wuhan", name_02 = "Alpha", name_03 = "Delta", name_04 =
    "Omicron1", name_05 = "Omicron2"),
  setSeed = F,
  seedValue = 100
)
}
\arguments{
\item{cases_ts}{time series of cases}

\item{time}{the number of observations in the time series}

\item{percent_genomic_surveillance_min}{the min percentage of cases that are sequenced}

\item{percent_genomic_surveillance_max}{the max percentage of cases that are sequenced}

\item{variant_start_time_fractions}{The time fraction for the start of var2-var5}

\item{variant_growth_rates}{the growth rate for each variant}

\item{variant_names}{the variant names}

\item{setSeed}{allows setting of the seed for sampling}

\item{seedValue}{set the seed value to be used if needed}
}
\value{
tibble output data
}
\description{
\code{synth_variants} 
Generates cases for five successive variants
}
